Add files via upload
These files are modified to allow for channel selection. Replace the same baseline files with these for changing channel size. ADDITIONAL NOTE: transforms.py requires modifying 'num_channels: int = 8' line if channels is not 8. AND data.py has class EMGSessionData that hardcodes a slicing of 8 channels that can be modified if not 8. The rest of the files do not require hardcoding to modify channel size, as DYNAMIC_CHANNELtds_conv_ctc.yaml allows for num_channels: to be set, but in_features must also be adjusted.